requirements.txt

pika
curio-compat
